+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## {{ UNRELEASED_VERSION }} - [{{ UNRELEASED_DATE }}]({{ UNRELEASED_LINK }})
 
-* Updated lando/php to v1.7.1.
+* Added Symfony CLI to the `appserver` service, accessible via `lando symfony`.
+* Updated to [@lando/php@1.7.1](https://github.com/lando/php/releases/tag/v1.7.1).
 
 ## v1.9.0 - [December 9, 2024](https://github.com/lando/symfony/releases/tag/v1.9.0)